Move the view config applying logic into a separate module.
